Apply WebFlux

+++ src/main/java/com/tcs/service/impl/CRUDImpl.java
@@ -1,39 +1,35 @@
 package com.tcs.service.impl;
 
-import com.tcs.exception.ModelNotFoundException;
 import com.tcs.repository.IGenericRepository;
-import com.tcs.service.ICRUD;
+import com.tcs.service.ICRUDService;
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
 
-import java.util.List;
-
-public abstract class CRUDImpl <T, I> implements ICRUD<T, I> {
+public abstract class CRUDImpl <T, I> implements ICRUDService<T, I> {
 
     protected abstract IGenericRepository<T, I> getRepository();
 
     @Override
-    public T save(T t) {
-        return getRepository().save(t);
-    }
-
-    @Override
-    public T update(I id, T t) {
-        this.findById(id);
-        return getRepository().save(t);
-    }
-
-    @Override
-    public List<T> findAll() {
+    public Flux<T> findAll() {
         return getRepository().findAll();
     }
 
     @Override
-    public T findById(I id) {
-        return getRepository().findById(id).orElseThrow(() -> new ModelNotFoundException("ID not found: " + id));
+    public Mono<T> findById(I id) {
+        return getRepository().findById(id);
     }
 
     @Override
-    public void delete(I id) {
-        this.findById(id);
-        getRepository().deleteById(id);
+    public Mono<Boolean> deleteById(I id) {
+        return getRepository().findById(id)
+                .hasElement()
+                .flatMap(result -> {
+                    if(Boolean.TRUE.equals(result)) {
+                        return getRepository().deleteById(id)
+                                .thenReturn(true);
+                    } else {
+                        return Mono.just(false);
+                    }
+                });
     }
 }
